:boom: Fix deprecation warnings
Removes override support for `showIndentGuide` and `showInvisibles`
because these have been deprecated and will require support for transient
config overrides as opposed to calling instance methods on an editor or view
like we used to be able to do. Once we support transient config values, we can
reimplement support for these overrides.